double not float, why say float then

+++ variadic_functions/3-print_all.c
@@ -47,7 +47,7 @@ void print_all(const char *format, ...)
 			printf("%d", va_arg(args, int));
 			break;
 		case ('f'):
-			printf("%f", va_arg(args, float));
+			printf("%f", va_arg(args, double));
 			break;
 		}
 		i++;
